backend/requirements.txt
django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.2.0
pillow>=10.0.0
coverage>=7.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.0.0
dj-database-url>=2.0.0
